requirements.txt
streamlit
pandas
pillow
python-dotenv
langchain
langchain-google-genai
google-generativeai
openai
chromadb
youtube-search-python
requests
